Split up LibXboxOne.Nand namespace
