+++ LibXboxOne.Tests/XbfsTests.cs
@@ -1,5 +1,6 @@
 using System.IO;
 using Xunit;
+using LibXboxOne.Nand;
 
 namespace LibXboxOne.Tests
 {
