+++ XBFSTool/Program.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using LibXboxOne;
+using LibXboxOne.Nand;
 using NDesk.Options;
 
 namespace XBFSTool
